secure-boot: check for manufacturing environment
This check is meant to be a "hey you didn't provision the system
correctly" type warning for the team building the system in a
manufacturing environment. There are other fundamental things that will
not work when a system is not secure. This check is just meant to be a
more user-friendly version so the team building the system can quickly
be notified that they've forgotten something.

There is no be-all "manufacturing environment" setting within OpenBMC so
utilize the one most often used within manufacturing to ensure a system
does not ship with any errors, the QuiesceOnHwError property.

Tested:
- Loaded application in QEMU, verified when QuiesceOnHwError was false
  and security settings were incorrect, no journal entry error was made.
- Verified when QuiesceOnHwError was true, and security settings were
  incorrect, that the "The system is not secure" entry was in the
  journal.

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Change-Id: I558e28593f30a939a39fb5cc5e201047c0b9e47b