only quiesce in reboot on host crash (#25)
This script was introduced to ensure a system properly halted in
scenarios where the host has crashed (obmc-host-crash@.target). There
was a loophole where the standard reboot attempt logic check was not
running in this specific path.

Unfortunately, that original change was not done quite right. This
script is actually called on all host reboots so we have run into
situations where the host is being put into quiesce before being given
one last try to boot. That last try in non-host crash scenarios is where
certain recovery actions (like trying alternate boot sources for some
devices) is done.

This commit modifies the script to ensure it only moves the host to the
quiesce state when the cause of the host reboot was a host crash.

Tested:
- Verified in normal reboot path, even if AttemptsLeft is 0, the reboot
  is still done
- Verified that if AttemptsLeft is 0 and reboot reason is HostCrash that
  host was put in quiesce

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>
Change-Id: I18e21cc33b3c38334ad760148a0754e2bd7d0780

Signed-off-by: Andrew Geissler <geissonator@yahoo.com>